# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Azure SQL TCO API
  version: 0.1.0
  description: >-
    Server-authoritative API for estimating AWS and on-premises SQL Server costs
    against Azure SQL Managed Instance. Results are estimates, not quotes.
servers:
  - url: /api/v1
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [Projects]
      security:
        - containerAppsAuth: []
      responses:
        '200':
          description: Projects owned by the current principal.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectSummary'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createProject
      tags: [Projects]
      security:
        - containerAppsAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditableProject'
      responses:
        '201':
          description: Created owner-scoped project.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/Problem'
  /projects/{project_id}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      operationId: getProject
      tags: [Projects]
      security:
        - containerAppsAuth: []
      responses:
        '200':
          description: Owner-scoped project.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/Problem'
    put:
      operationId: updateProject
      tags: [Projects]
      security:
        - containerAppsAuth: []
      parameters:
        - name: If-Match
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditableProject'
      responses:
        '200':
          description: Updated owner-scoped project.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteProject
      tags: [Projects]
      security:
        - containerAppsAuth: []
      responses:
        '204':
          description: Project deleted immediately.
        default:
          $ref: '#/components/responses/Problem'
components:
  securitySchemes:
    containerAppsAuth:
      type: openIdConnect
      openIdConnectUrl: https://login.microsoftonline.com/common/v2.0/.well-known/openid-configuration
      description: Authentication is enforced by Azure Container Apps built-in authentication.
  parameters:
...
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Problem:
      description: RFC 9457 Problem Details response.
      headers:
        Retry-After:
          description: Seconds to wait when the request quota is exceeded.
          schema:
            type: string
        ETag:
          description: Current project ETag after a failed precondition.
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Decimal:
      type: string
      pattern: '^-?(0|[1-9][0-9]*)(\.[0-9]+)?$'
      examples: ['8760', '0.10', '12345.67']
...
    PurchaseOptionKey:
      type: string
      enum: [payg, ahb, one-year, ahbone-year, three-year, ahbthree-year, sv-one-year, ahbsv-one-year]
...
    CalculationRevision:
      type: object
      required: [formula_version, aws_snapshot_id, azure_snapshot_id, resource_results, portfolio_totals, warnings]
      properties:
        formula_version:
          type: string
        aws_snapshot_id:
          type: [string, 'null']
        azure_snapshot_id:
          type: [string, 'null']
        resource_results:
          type: array
          items:
            $ref: '#/components/schemas/ResourceCalculation'
        portfolio_totals:
          $ref: '#/components/schemas/PortfolioTotals'
        warnings:
          type: array
          items:
            type: string
    ResourceCalculation:
      type: object
      required:
        - resource_id
        - mapping_status
        - aws_pricing_status
        - azure_pricing_status
        - target_selection
        - source_costs
        - azure_costs
        - savings
        - explanation_steps
        - unresolved_components
      properties:
        resource_id:
          type: string
          format: uuid
        mapping_status:
          type: [string, 'null']
          enum: [mapped, no_mapping, null]
        aws_pricing_status:
          $ref: '#/components/schemas/PricingStatus'
        azure_pricing_status:
          $ref: '#/components/schemas/PricingStatus'
        target_selection:
          oneOf:
            - $ref: '#/components/schemas/TargetSelection'
            - type: 'null'
        source_costs:
          oneOf:
            - $ref: '#/components/schemas/SourceCostBreakdown'
            - type: 'null'
        azure_costs:
          oneOf:
            - $ref: '#/components/schemas/AzureCostBreakdown'
            - type: 'null'
        savings:
          oneOf:
            - $ref: '#/components/schemas/SavingsBreakdown'
            - type: 'null'
        explanation_steps:
          type: array
          items:
            $ref: '#/components/schemas/ExplanationStep'
        unresolved_components:
          type: array
          items:
            $ref: '#/components/schemas/UnresolvedComponent'
    PricingStatus:
      type: string
      enum: [fresh, cached, stale, unavailable, not_required]
    PortfolioTotals:
      type: object
      required:
        - aws_all_rows_total
        - aws_mapped_rows_total
        - azure_mapped_rows_total
        - required_portfolio_adjustment
        - selected_parity_adjustment
        - portfolio_after_selected_parity
        - portfolio_difference
        - comparable_resource_count
        - no_mapping_resource_count
        - price_unavailable_resource_count
      properties:
        aws_all_rows_total:
          oneOf:
            - $ref: '#/components/schemas/Decimal'
            - type: 'null'
        aws_mapped_rows_total:
          $ref: '#/components/schemas/Decimal'
        azure_mapped_rows_total:
          $ref: '#/components/schemas/Decimal'
        required_portfolio_adjustment:
          $ref: '#/components/schemas/Decimal'
        selected_parity_adjustment:
          $ref: '#/components/schemas/Decimal'
        portfolio_after_selected_parity:
          $ref: '#/components/schemas/Decimal'
        portfolio_difference:
          $ref: '#/components/schemas/Decimal'
        comparable_resource_count:
          type: integer
          minimum: 0
        no_mapping_resource_count:
          type: integer
          minimum: 0
        price_unavailable_resource_count:
          type: integer
          minimum: 0
    SourceCostBreakdown:
      type: object
      required: [compute_gross, compute_net, license_gross, license_net, storage_gross, storage_net, hardware_annual, electricity_annual, total]
      properties:
        compute_gross:
          $ref: '#/components/schemas/Decimal'
        compute_net:
          $ref: '#/components/schemas/Decimal'
        license_gross:
          $ref: '#/components/schemas/Decimal'
        license_net:
          $ref: '#/components/schemas/Decimal'
        storage_gross:
          $ref: '#/components/schemas/Decimal'
        storage_net:
          $ref: '#/components/schemas/Decimal'
        hardware_annual:
          $ref: '#/components/schemas/Decimal'
        electricity_annual:
          $ref: '#/components/schemas/Decimal'
        total:
          $ref: '#/components/schemas/Decimal'
    AzureCostBreakdown:
      type: object
      required: [compute_gross, additional_ram_gb, additional_ram_gross, compute_plus_ram_net, license_gross, license_net, storage_gross, storage_net, total_before_parity]
      properties:
        compute_gross:
          $ref: '#/components/schemas/Decimal'
        additional_ram_gb:
          $ref: '#/components/schemas/Decimal'
        additional_ram_gross:
          $ref: '#/components/schemas/Decimal'
        compute_plus_ram_net:
          $ref: '#/components/schemas/Decimal'
        license_gross:
          $ref: '#/components/schemas/Decimal'
        license_net:
          $ref: '#/components/schemas/Decimal'
        storage_gross:
          $ref: '#/components/schemas/Decimal'
        storage_net:
          $ref: '#/components/schemas/Decimal'
        total_before_parity:
          $ref: '#/components/schemas/Decimal'
    SavingsBreakdown:
      type: object
      required: [compute_savings, license_savings, storage_savings, total_savings, required_adjustment, selected_adjustment, azure_after_selected_parity, difference]
      properties:
        compute_savings:
          $ref: '#/components/schemas/Decimal'
        license_savings:
          $ref: '#/components/schemas/Decimal'
        storage_savings:
          $ref: '#/components/schemas/Decimal'
        total_savings:
          $ref: '#/components/schemas/Decimal'
        required_adjustment:
          $ref: '#/components/schemas/Decimal'
        selected_adjustment:
          $ref: '#/components/schemas/Decimal'
        azure_after_selected_parity:
          $ref: '#/components/schemas/Decimal'
        difference:
          $ref: '#/components/schemas/Decimal'
    TargetSelection:
      type: object
      required: [mapping_status, requested_tier, nggp_iops_limit, selected, candidates, outcome_reasons, storage_escalation]
      properties:
        mapping_status:
          type: string
          enum: [mapped, no_mapping]
        requested_tier:
          $ref: '#/components/schemas/ServiceTier'
        nggp_iops_limit:
          type: integer
          minimum: 0
        selected:
          oneOf:
            - $ref: '#/components/schemas/SelectedTarget'
            - type: 'null'
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/CandidateEvaluation'
        outcome_reasons:
          type: array
          items:
            $ref: '#/components/schemas/SelectionReason'
        storage_escalation:
          oneOf:
            - $ref: '#/components/schemas/StorageEscalation'
            - type: 'null'
    ServiceTier:
      type: string
      enum: [next_generation_general_purpose, business_critical]
    SelectedTarget:
      type: object
      required: [configuration_key, azure_region, service_tier, hardware_family, vcores, included_memory_gb, selected_memory_gb, additional_memory_gb, storage_architecture, maximum_storage_gb]
      properties:
        configuration_key:
          type: string
        azure_region:
          type: string
        service_tier:
          $ref: '#/components/schemas/ServiceTier'
        hardware_family:
          type: string
        vcores:
          type: integer
          minimum: 1
        included_memory_gb:
          $ref: '#/components/schemas/Decimal'
        selected_memory_gb:
          $ref: '#/components/schemas/Decimal'
        additional_memory_gb:
          $ref: '#/components/schemas/Decimal'
        storage_architecture:
          type: string
        maximum_storage_gb:
          oneOf:
            - $ref: '#/components/schemas/Decimal'
            - type: 'null'
    CandidateEvaluation:
      type: object
      required: [configuration_key, vcores, selected_memory_gb, eligible, rejection_reasons]
      properties:
        configuration_key:
          type: string
        vcores:
          type: integer
        selected_memory_gb:
          oneOf:
            - $ref: '#/components/schemas/Decimal'
            - type: 'null'
        eligible:
          type: boolean
        rejection_reasons:
          type: array
          items:
            $ref: '#/components/schemas/SelectionReason'
    SelectionReason:
      type: object
      required: [code, detail]
      properties:
        code:
          type: string
          enum: [region_unavailable, tier_unavailable, zone_redundancy_unsupported, hardware_family_unsupported, insufficient_vcores, insufficient_memory, insufficient_storage]
        detail:
          type: string
    StorageEscalation:
      type: object
      required: [rejected_configuration_key, rejected_maximum_storage_gb, selected_configuration_key]
      properties:
        rejected_configuration_key:
          type: string
        rejected_maximum_storage_gb:
          $ref: '#/components/schemas/Decimal'
        selected_configuration_key:
          type: string
    ExplanationStep:
      type: object
      required: [code, message, values]
      properties:
        code:
          type: string
        message:
          type: string
        values:
          type: object
          additionalProperties:
            type: string
    UnresolvedComponent:
      type: object
      required: [provider, code, message]
      properties:
        provider:
          type: [string, 'null']
          enum: [aws, azure, null]
        code:
          type: string
        message:
          type: string
    EditableProject:
      allOf:
        - $ref: '#/components/schemas/EditableProjectFields'
      unevaluatedProperties: false
    EditableProjectFields:
      type: object
      required: [name, settings, resources, aws_price_snapshot_id, azure_price_snapshot_id]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: [string, 'null']
          maxLength: 500
        settings:
          $ref: '#/components/schemas/ProjectSettings'
        resources:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/Resource'
        aws_price_snapshot_id:
          type: [string, 'null']
        azure_price_snapshot_id:
          type: [string, 'null']
    ProjectSettings:
      type: object
      additionalProperties: false
      required:
        - project_type
        - aws_region
        - azure_region
        - currency
        - source_compute_discount
        - source_license_discount
        - source_storage_discount
        - azure_compute_discount
        - azure_license_discount
        - azure_storage_discount
        - selected_parity_adjustment
        - default_annual_hours
        - default_mi_purchase_option
      properties:
        project_type:
          type: string
          enum: [ec2, rds, on_prem]
        aws_region:
          type: [string, 'null']
        azure_region:
          type: string
        currency:
          type: string
          const: USD
        source_compute_discount:
          $ref: '#/components/schemas/Decimal'
        source_license_discount:
          $ref: '#/components/schemas/Decimal'
        source_storage_discount:
          $ref: '#/components/schemas/Decimal'
        azure_compute_discount:
          $ref: '#/components/schemas/Decimal'
        azure_license_discount:
          $ref: '#/components/schemas/Decimal'
        azure_storage_discount:
          $ref: '#/components/schemas/Decimal'
        selected_parity_adjustment:
          $ref: '#/components/schemas/Decimal'
        default_annual_hours:
          $ref: '#/components/schemas/Decimal'
        default_mi_purchase_option:
          $ref: '#/components/schemas/PurchaseOptionKey'
        enterprise_license_sa_usd_per_two_core_pack:
          oneOf:
            - $ref: '#/components/schemas/Decimal'
            - type: 'null'
        standard_license_sa_usd_per_two_core_pack:
          oneOf:
            - $ref: '#/components/schemas/Decimal'
            - type: 'null'
        remaining_coverage_months:
          type: [integer, 'null']
          enum: [12, 24, 36, null]
        electricity_rate_usd_per_kwh:
          oneOf:
            - $ref: '#/components/schemas/Decimal'
            - type: 'null'
    Resource:
      oneOf:
        - $ref: '#/components/schemas/Ec2Resource'
        - $ref: '#/components/schemas/RdsResource'
        - $ref: '#/components/schemas/OnPremResource'
      discriminator:
        propertyName: source_type
    SharedResourceProperties:
      type: object
      required:
        - id
        - workload_name
        - quantity
        - sql_edition
        - license_basis
        - sql_data_gb_per_instance
        - source_ram_gb_per_instance
        - annual_hours_per_instance
        - mi_purchase_option
      properties:
        id:
          type: string
          format: uuid
        workload_name:
          type: string
          minLength: 1
          maxLength: 160
        quantity:
          type: integer
          minimum: 1
          maximum: 10000
        sql_edition:
          type: string
          enum: [standard, enterprise]
        license_basis:
          type: string
          enum: [license_included, byol]
        sql_data_gb_per_instance:
          $ref: '#/components/schemas/Decimal'
        source_ram_gb_per_instance:
          $ref: '#/components/schemas/Decimal'
        annual_hours_per_instance:
          $ref: '#/components/schemas/Decimal'
        mi_purchase_option:
          $ref: '#/components/schemas/PurchaseOptionKey'
    Ec2Resource:
      allOf:
        - $ref: '#/components/schemas/SharedResourceProperties'
        - type: object
          required: [source_type, instance_type, volumes]
          properties:
            source_type:
              type: string
              const: ec2
            instance_type:
              type: string
            volumes:
              type: array
              minItems: 1
              maxItems: 50
              items:
                $ref: '#/components/schemas/EbsVolume'
    EbsVolume:
      type: object
      required: [id, label, aws_volume_id, volume_type, capacity_gb, provisioned_iops, throughput_mibps]
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
          minLength: 1
          maxLength: 80
        aws_volume_id:
          type: [string, 'null']
          maxLength: 128
        volume_type:
          type: string
          enum: [gp3, io2, ephemeral]
        capacity_gb:
          $ref: '#/components/schemas/Decimal'
        provisioned_iops:
          type: [integer, 'null']
          minimum: 0
        throughput_mibps:
          oneOf:
            - $ref: '#/components/schemas/Decimal'
            - type: 'null'
    RdsResource:
      allOf:
        - $ref: '#/components/schemas/SharedResourceProperties'
        - type: object
          required: [source_type, instance_type, deployment, commercial_term, storage_class, source_max_iops]
          properties:
            source_type:
              type: string
              const: rds
            instance_type:
              type: string
            deployment:
              type: string
              enum: [single_az, multi_az]
            commercial_term:
              type: string
            storage_class:
              type: string
            source_max_iops:
              type: integer
              minimum: 0
              maximum: 1000000000
    OnPremResource:
      allOf:
        - $ref: '#/components/schemas/SharedResourceProperties'
        - type: object
          required: [source_type, source_vcpu, licensable_cores, source_max_iops, hardware_capex_usd, depreciation_years]
          properties:
            source_type:
              type: string
              const: on_prem
            source_vcpu:
              type: integer
              minimum: 1
              maximum: 100000
            licensable_cores:
              type: integer
              minimum: 1
              maximum: 100000
            source_max_iops:
              type: integer
              minimum: 0
              maximum: 1000000000
            hardware_capex_usd:
              $ref: '#/components/schemas/Decimal'
            depreciation_years:
              $ref: '#/components/schemas/Decimal'
            average_power_kw_override:
              oneOf:
                - $ref: '#/components/schemas/Decimal'
                - type: 'null'
    Project:
      allOf:
        - $ref: '#/components/schemas/EditableProjectFields'
        - type: object
          required: [id, created_at, updated_at, formula_version, schema_version, latest_calculation_revision]
          properties:
            id:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
            formula_version:
              type: string
            schema_version:
              type: string
            latest_calculation_revision:
              oneOf:
                - $ref: '#/components/schemas/CalculationRevision'
                - type: 'null'
      unevaluatedProperties: false
    ProjectSummary:
      type: object
      required: [id, name, project_type, modified_at, source_region, azure_region, resource_count, source_annual_total, azure_annual_total]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        project_type:
          type: string
          enum: [ec2, rds, on_prem]
        modified_at:
          type: string
          format: date-time
        source_region:
          type: [string, 'null']
        azure_region:
          type: string
        resource_count:
          type: integer
        source_annual_total:
          oneOf:
            - $ref: '#/components/schemas/Decimal'
            - type: 'null'
        azure_annual_total:
          oneOf:
            - $ref: '#/components/schemas/Decimal'
            - type: 'null'
    Problem:
      type: object
      required: [type, title, status, detail, instance, request_id]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
        request_id:
          type: string
        errors:
          type: array
          items:
            type: object
            required: [pointer, code, message]
            properties:
              pointer:
                type: string
              code:
                type: string
              message:
                type: string